I dont think this would work.
I want to make the artboard section apear after 5 seconds of page load.
The artboard is inside main section.
I was trying to make another empty main section appear over the actual
main section. Then give the main section a higher z index after 5
seconds.
I cant get it to work.

I could dive deep into why it is not working... this is what i have been
doing since the start of this project and i have gotten it to work
everysingle time though with a lot of dertermination and strong will.

how ever, i thing there is an alternative to this solution
and i am not really sure that vanilla js can re-render the screen.
I think that is what i am trying to do. And it is what i hear React js
does.