Crazy solution to nextBtn
My goal was to check if count equals or greater than array.lenth
then display a message and exit the function

but this wasnt working. I knew i had to look for the issue.

In the process of looking on my own, i found something.
i know it is exactly what i was looking for.

when ths nextBtn is clicked
- increase count
- check if coount >= arraylenght
- - if yes, display a text on page and reduce decrease count
- - if no, run the clearboard and displays

what this does is make the next be in lool of 4 and 3, if araylenth is 4

so if the last elemt in the array is array[3]
and u clicked the nextbtn again
3 turns 4
4 meets the condition
message is displayed and count reduced (4 reduced to 3)
and function ends
you clicked next again and steps repeats

I think i have to look for another way..i am sure there is
or atleats see how others handle this scenerio