What is 'this' keyword
this keyword is use by a method to indicate the object the fn is
refering to.

method is a function that is a property of an object.
object is a key-value paired data type that uses strings as property
name
array is also an object but uses number as property name